requirements.txt
aniso8601==8.0.0
argon2-cffi==19.1.0
astroid==2.3.0
certifi==2019.9.11
cffi==1.12.3
Click==7.0
colorama==0.4.1
Flask==1.1.1
Flask-Login==0.4.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.19
pylint==2.4.1
pymongo==3.9.0
python-engineio==3.11.1
python-socketio==4.4.0
pytz==2019.3
six==1.12.0
typed-ast==1.4.0
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
